Refactor: Split TypeBaseExtensions into partial classes
