+++ src/ModelFramework.Objects.Tests/Extensions/TypeBaseExtensionsTests.General.cs
@@ -1,13 +1,16 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using FluentAssertions;
 using ModelFramework.Objects.Builders;
+using ModelFramework.Objects.Contracts;
 using ModelFramework.Objects.Extensions;
+using Moq;
 using Xunit;
 
 namespace ModelFramework.Objects.Tests.Extensions
 {
     [ExcludeFromCodeCoverage]
-    public class TypeBaseExtensionsTests
+    public partial class TypeBaseExtensionsTests
     {
         [Fact]
         public void GetInheritedClasses_Returns_Empty_String_For_Interface_When_No_Interfaces_Are_Defined()
@@ -94,6 +97,43 @@ namespace ModelFramework.Objects.Tests.Extensions
 
             // Assert
             actual.Should().Be(" : BaseClass");
+        }
+
+        [Fact]
+        public void GetContainerType_Returns_Correct_Value_For_Class()
+        {
+            // Arrange
+            var sut = new ClassBuilder().WithName("Test").Build();
+
+            // Act
+            var actual = sut.GetContainerType();
+
+            // Assert
+            actual.Should().Be("class");
+        }
+
+        [Fact]
+        public void GetContainerType_Returns_Correct_Value_For_Interface()
+        {
+            // Arrange
+            var sut = new InterfaceBuilder().WithName("Test").Build();
+
+            // Act
+            var actual = sut.GetContainerType();
+
+            // Assert
+            actual.Should().Be("interface");
+        }
+
+        [Fact]
+        public void GetContainerType_Throws_On_Unknown_Type()
+        {
+            // Arrange
+            var sut = new Mock<ITypeBase>().Object;
+            var action = new Action(() => sut.GetContainerType());
+
+            // Act & Assert
+            action.Should().Throw<InvalidOperationException>().WithMessage("Unknown container type: [Castle.Proxies.ITypeBaseProxy]");
         }
     }
 }
